fix(tests): remove Value(nLL) ambiguity on Linux
On Linux x86-64 int64_t is long, so a long long literal argument
converts equally well to the int64_t and double Value constructors and
gcc rejects the call as ambiguous; on macOS int64_t is long long and
the exact match hides this. All 148 Value(<n>LL) sites across three
test files now use Value(int64_t(<n>)), matching the style used
everywhere else. Verified with a full gcc-13 Release build and
382/382 test run on linux/amd64 under Docker with -mavx2 — which also
executed the AVX2 join path with the whole suite.

+++ tests/test_binder.cc
@@ -286,10 +286,10 @@ TEST(SelfJoin, IsNullResolvesJoinSideColumn) {
         {"k", TypeId::INT, 0}, {"v", TypeId::INT, 0},
         {"k", TypeId::INT, 1}, {"w", TypeId::INT, 1}});
 
-    std::vector<Row> from_rows = {{Value(1LL), Value(10LL)}};
+    std::vector<Row> from_rows = {{Value(int64_t(1)), Value(int64_t(10))}};
     std::vector<Row> join_rows = {
-        {Value(1LL), Value::null()},   // w is NULL
-        {Value(1LL), Value(99LL)},     // w is 99
+        {Value(int64_t(1)), Value::null()},   // w is NULL
+        {Value(int64_t(1)), Value(int64_t(99))},     // w is 99
     };
     auto from_scan = std::make_unique<SeqScanNode>("a", from_rows, from_schema);
     auto join_scan = std::make_unique<SeqScanNode>("b", join_rows, join_schema);
